feat: Add flexible API pagination and update dependencies
Introduces support for multiple pagination strategies (page number, cursor, limit/offset) across all list endpoints, with new query parameters and documentation. Refactors pagination logic, updates OpenAPI schema, and improves serializer and view handling for pagination. Removes unused blog comment endpoint, enhances product review serialization, and updates various factories for unique image filenames. Upgrades several dependencies and MeiliSearch version, and adds a management command to clear MeiliSearch indexes.

+++ product/serializers/image.py
@@ -64,10 +64,10 @@ class ProductImageSerializer(
             return obj.image.url
         return ""
 
-    def get_image_size_kb(self, obj: ProductImage) -> int:
+    def get_image_size_kb(self, obj: ProductImage) -> float:
         if obj.image and hasattr(obj.image, "size"):
             return round(obj.image.size / 1024, 2)
-        return 0
+        return 0.0
 
     def get_alt_text(self, obj: ProductImage) -> str:
         title = obj.safe_translation_getter("title", any_language=True)
